refactor: app and page structure
