+++ app/pages/release/[slug].vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-const route = useRoute();
+const route = useRoute('release-slug');
 const client = useSupabaseClient();
 
 if (!route.params.slug)
